Use a NullFlash object when the session store is disabled
The module `EtagWithFlash` (introduced in #26250) uses flash. However,
when the session store is not set (or disabled), it will raise an
error because the flash middleware is included only if a session store
is present.

This patch includes the middleware even when the session store is
disabled, although the flash module will return a null object to
prevent breaks.