Change core_ext cherry picking example.
`Object#blank?` is used as an example in this guide of a core extension
that is not loaded by default in Active Support, and how to load it
manually.  However it now _is_ loaded by default.

This change gives a different example (`Hash#with_indifferent_access`)
that will hopefully be less likely to be pulled into the default
requires in future.